testbed: Remove trailing dot from path references
Issue: VLAN tests fail to configure the vlan on the lan side because a VLANPort can not be created

Cause: Trailing dots are no longer allowed in the path references. TR181-bridging v2.8.13, v2.8.10 and v2.8.12 include changes to remove the trailing dot from path references. This is done to be inline with the tr181 spec

Fix: Remove the trailing dots from the command creating the VLANPort

Signed-off-by: Toon Caeyers <toon.caeyers_ext@softathome.com>