PROVISORY: testbed config: Increase tr69InformTimeout to 150s
After  upstepping multiple components, including tr069-manager, we see an increase in failing CDRouter tests failing because the inform is not sent in time for the test.

This issue is mostly seen on Freedom boards as they are booted up quicker we see the following scenario:
* Board receives a 10.0.0.2 address
* It starts trying to send out the tr069 inform, failing to do so, increasing the retry timeout every time
* After a while the board gets the new (correct) IP address from the CDRouter.
* The retry timeout is now big enough that the inform is not send in time, test times out  before the inform

If this test fails, none of the CDRouter tests in this job run. By increasing the InformTimeout for now,
this test should no longer fail and we should be able to run the CDRouter tests again.

References: PPW-544
Signed-off-by: Toon Caeyers <toon.caeyers_ext@softathome.com>