ci: cram: obuspa: add tr181-bridging MACEvent
A generic event in Bridging that is emitted when a device is
detected behind a port is needed. Since the MAC address
uniquely identifies a device, a "MACEvent!"" will be
emitted when a device is detected behind a certain port.

This new event is defined in the Bridging data model and is
now also seen by the obuspa test. Therefore the test needs to
be updated with these new expectations.

References: PPW-700
Signed-off-by: Sarra Kacem <sarra.kacem_ext@softathome.com>